pyproject.toml
[project]
name = "todoaiagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Lukas Hamm",email = "lukas@lukashamm.dev"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "streamlit (>=1.49.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-ai-slim[mistral] (>=1.0.17,<2.0.0)",
    "speechrecognition (>=3.14.3,<4.0.0)",
    "langchain-mistralai (>=1.0.1,<2.0.0)",
    "langchain (>=1.0.2,<2.0.0)",
    "langgraph (>=1.0.2,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "todoaiagent", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"
mypy = "^1.18.2"
pre-commit = "^4.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
